// ui/src/types/generated/ScanSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MediaFile } from "./MediaFile";
import type { PlayAllInfo } from "./PlayAllInfo";
import type { ScanWarning } from "./ScanWarning";

/**
 * Everything the Confirm show screen needs about a scanned folder.
 */
export type ScanSummary = { 
/**
 * The folder that was scanned.
 */
folder: string, 
/**
 * Every video file found, sorted by [`FileId`].
 */
files: Array<MediaFile>, 
/**
 * The detected play-all title, if any.
 */
playAll: PlayAllInfo | null, 
/**
 * Number of [`FileRole::Candidate`] files.
 */
candidateCount: number, 
/**
 * Show name guessed from the folder or disc name, used as the initial search text.
 */
showGuess: string | null, 
/**
 * Problems the user should see before identifying.
 */
warnings: Array<ScanWarning>, };